pyproject.toml
[project]
# TODO: split to template and cookiecutter .toml
name = "drf-project-template"
version = "0.0.1"
description = "Template for creating django rest framework project."
authors = [
    {name = "Tvo-Po", email = "maxwellgleb@gmail.com"},
]
dependencies = [
    "django~=4.2.0",
    "djangorestframework~=3.14.0",
    "cookiecutter~=2.5.0",
    "django-environ~=0.11.2",
    "django-cors-headers~=4.3.1",
    "django-filter~=23.5",
    "drf-spectacular~=0.26.5",
]
requires-python = "~=3.11.0"
readme = "README.md"
license = {text = "MIT"}

